skill-packages/hushclaw-skill-pptx/requirements.txt
python-pptx>=0.6.23
